Switch to faye/websocket client for proxy support
The interface should be exactly the same but the error thrown on SSL
verification failure is now WebsocketError instead of an OpenSSL error.
This error will also now be thrown from any of the methods that execute
a job if it occurs before authentication has completed.  Previously
those methods would throw a RuntimeException due to authentication not
completing within the 5 second timeout due to an error.