allow specifying SSL args ca_file, client_cert, client_key
In our environment, we have a deployment of
https://github.com/signalfx/pops.

To send metrics to this POPS instance, we need to specify the CA file
(because the server cert is issued by our internal CA).

We also need to specify the client certificate/key (because our POPS
instance expects clients to authenticate).

The SignalFxClient's current use of RestClient:Request doesn't allow
specifying these options at the moment, so this commit adds them.